+++ Source/AST/Nodes/BaseNode.swift
@@ -7,10 +7,29 @@
 //
 
 import Foundation
+import libcmark
 
 public class BaseNode: Node {
     
     public let cmarkNode: CMarkNode
+    
+    public private(set) lazy var childen: [Node] = {
+        var result: [Node] = []
+        var child = cmark_node_first_child(cmarkNode)
+        
+        while let raw = child {
+            
+            guard let node = raw.wrap() else {
+                assertionFailure("Couldn't wrap node of type: \(raw.type)")
+                continue
+            }
+            
+            result.append(node)
+            child = cmark_node_next(child)
+        }
+        
+        return result
+    }()
     
     init(cmarkNode: CMarkNode) {
         self.cmarkNode = cmarkNode
