+++ Source/AST/Nodes/Node.swift
@@ -8,8 +8,6 @@
 import Foundation
 import libcmark
 
-public typealias CMarkNode = UnsafeMutablePointer<cmark_node>
-
 /// A node is a wrapper of a raw `CMarkNode` belonging to the abstract syntax tree
 /// generated by cmark.
 ///
@@ -21,34 +19,17 @@ public protocol Node {
     var childen: [Node] { get }
 }
 
-extension Node {
-    /// The children of this node.
-    public var childen: [Node] {
-        var result: [Node] = []
-        var child = cmark_node_first_child(cmarkNode)
-        
-        while let raw = child {
-            
-            guard let node = raw.wrap() else {
-                assertionFailure("Couldn't wrap node of type: \(raw.type)")
-                continue
-            }
-            
-            result.append(node)
-            child = cmark_node_next(child)
-        }
-        
-        return result
-    }
-    
+public extension Node {
     /// True iff the node has a sibling that succeeds it.
-    public var hasSuccessor: Bool {
+    var hasSuccessor: Bool {
         return cmark_node_next(cmarkNode) != nil
     }
 }
 
 
 // MARK: - Helper extensions
+
+public typealias CMarkNode = UnsafeMutablePointer<cmark_node>
 
 public extension UnsafeMutablePointer where Pointee == cmark_node {
